+++ lib/waterdrop/patches/rdkafka/producer.rb
@@ -10,7 +10,15 @@ module WaterDrop
         # Adds a method that allows us to get the native kafka producer name
         # @return [String] producer instance name
         def name
-          ::Rdkafka::Bindings.rd_kafka_name(@native_kafka)
+          unless @_native
+            version = ::Gem::Version.new(::Rdkafka::VERSION)
+            change = ::Gem::Version.new('0.12.0')
+            # 0.12.0 changed how the native producer client reference works.
+            # This code supports both older and newer versions of rdkafka
+            @_native = version >= change ? @client.native : @native_kafka
+          end
+
+          ::Rdkafka::Bindings.rd_kafka_name(@_native)
         end
       end
     end